armv8: make multientry conditional
While ARMv8 SoC will enter U-Boot both on a master and a secondary
(slave) CPU, this is not the common behaviour on ARMv8 platforms.
The norm is that U-Boot is entered from the master CPU only,
while the other CPUs are kept in WFI (wait for interrupt) state.

The code determining which CPU we are running on is using the
MPIDR register, but the definition of that register varies with
platform to some extent, and handling multi-cluster platforms
(such as the Juno) will become cumbersome. It is better to only
enable the multiple entry code on machines that actually need
it and disable it by default.

Make the single entry default and add a special
ARMV8_MULTIENTRY KConfig option to be used by the
platforms that need multientry and set it for ARMv8 Soc.
make it conditional in the generic start.S code.

This makes the Juno platform start U-Boot properly.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: cwz <cwz@rock-chips.com>