rk3399: clock: add reset interface for PCIe
w/o the basic framework of reset controller, it becomes
difficult for us to consolidate the reset ops inside the
box. It is also pointless to introduce dt-bindings parse
helper function to get them. Hardcode it in the rk3399
clock file and expose the API which keeps the model as-is.

Change-Id: Iec9026ff8e708f004a32b5a9b1f02961d3ca7106
Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>