tool: loaderimage: limit the uboot size and backup number for NVMe boot
Generally speaking, if supporting NVMe boot, we need a extra spi-nor
to flash and load uboot.img. Given that the low efficiency of bootrom's
spi code, let's try to cut down the size of uboot.img and the backup
number to speedup the flow. Note that it's mandatory to limit it now
by the spi_fw_tools so that let's improve it.

Change-Id: I00c09fc72ddcc8e3c7bdd838c4b06927ae55b32c
Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>