rk33plat: support usb bypass uart for rk3399
rk3399 Type-C0 USB 2.0 DP/DM can be bypassed to UART, it's
useful for the rk3399 platforms without UART interface to
print log via USB interface.

With this patch, the user still can't use this bypass function.
It needs to add special param in the parameter file as following:

USBUART: enable

And it also needs a special USB cable integrated with an USB to
UART chip.

Change-Id: I2c2f029c4203bc4a05708268dfaa967b3054e7fc
Signed-off-by: Wu Liang feng <wulf@rock-chips.com>