rockchip: video: reserve memory for fb without bind video driver
The video DM need to bind video driver before relocate in order to
get the size need by video driver, which is fixed in rockchip platform,
we can get it without bind the driver.

With this patch, we don't need to enable the video/display nodes
in pre-reloc, and we can get all these nodes from kernel dtb.

Change-Id: I02d86b69e8c10bbf47e6b421c41b6dae20667a33
Signed-off-by: Kever Yang <kever.yang@rock-chips.com>