drm/rockchip: dw-hdmi: sort mode list before select hdmi display mode
If preferred mode defined by EDID is not in the resolution white list,
the original way is to set the EDID preferred mode, that is not match
the mode defined in Android. Now sort edid mode list first in this
case the same as do in android to fix this problam.

Change-Id: Iadab882bdcaa81a713d29a8f2895b65cdd2e848f
Signed-off-by: Huicong Xu <xhc@rock-chips.com>