pwm: rockchip: Do pwm pinctrl setting after pwm enabled
If the PWM pinctrl uses default state, the iomux setting will
be done at probe, the PWM may not be enabled at this moment.
It will make PWM into an intermediate state, destroy the default
hardware state, the PWM is not ready for work yet. So it is better
for doing PWM pinctrl setting after PWM enabled.

Change-Id: I8b700a07083b9a6798303fc5a5ed9f29eb42c47e
Signed-off-by: David Wu <david.wu@rock-chips.com>
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>