pwm: rk_pwm: Make PWM pinctrl setting after PWM enabled
If the PWM pinctrl uses default state, in the probe will be set,
at this time the PWM may not be enabled. There is a period of
time for PWM enabled, it will make PWM in an intermediate state,
destroy the hardware defaults, so it will be better PWM pinctrl
setting after PWM enabled.

Change-Id: Iab087c03218f223b2589c3a1d4af62a0a251950b
Signed-off-by: David Wu <david.wu@rock-chips.com>