drm/rockchip: Unbind inno hdmi phy and hdmi connector
In some platforms, the normal operation of different connectors need
the same phy. Such as rk322x and rk322xh, Hdmi and tve get clk from
the same public phy(inno hdmi phy).
In the original display frame, the public phy is bound with specific
connector. For example, inno hdmi phy and hdmi connector are
bound together. It will lead to a problem that if hdmi is disabled
hdmi phy clk can't be set and tve can't get clk.
To solve this problem, inno hdmi phy and hdmi connector should not
be bound together. The first initialized one of connectors which using
this public phy will initialize this phy and saved phy data, then
other connectors which use that phy can get it directly. Even if one of
these connector is disabled, others can work normally.

Change-Id: I4eb08340617cb0e21a29673e509e34dcedb6ccd6
Signed-off-by: Algea Cao <algea.cao@rock-chips.com>