Set right offset for dtb file
Set right offset for dtb file if read from second pos in android recovery.img.
The right offset is used to locate the file and check the header but when added to the file list the wrong base address is used. The file points to the image header not to the dtb on later usage.