video/drm: lvds: add support for dual lvds connector
There are two models of dual-lvds:
1. dual-channel-lvds: This mode has one panel with two separate
LVDS connector outputs for even and odd pixel.

2. dual-lvds: This mode has two panel, outputting the left content
and the right content respectively.

These two modes are regarded as the connectors in the split-mode,
just like dp/edp/hdmi.

In order to be consistent with the kernel, an additional DTS bool
attribute of "dual-channel" is added to help VOP determine whether
there are multiple LVDS connectors.

Signed-off-by: Chaoyi Chen <chaoyi.chen@rock-chips.com>
Change-Id: I1d3b3cfa20419db5fe089f89f8a3b91c40903e97