video/drm: lvds: add support for secondary lvds probe
Uses a method similar to the kernel to probe dual lvds.

Note that there is no defer mechanism in uboot, so these
operations need to be performed when the first device is loaded.

The way of sharing data between the two connectors is
also different, but the overall operation is similar to the kernel.

Signed-off-by: Chaoyi Chen <chaoyi.chen@rock-chips.com>
Change-Id: I77dfff8dfe3cb37490b346fa6cd26cf189d08ae7