video/drm: vop2: check the hot plug capability for plane mask assignment only for RK3566
For RK3566, the main planes should be enabled before the mirror planes.
The devices that support hot plug may be disconnected initially, so we
assign the main planes to the first device that does not support hot
plug, in order to ensure that the mirror planes are not enabled first.

For the platforms other than RK3566, we assign the plane mask of VPx
based on the &vop2_data.plane_mask[active_vp_num][x].

Change-Id: I7a5cc967623bf4aeb7d4bb9c5b0a36a288c30fc8
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>
(cherry picked from commit 7a29012829e857d1aaef78279325f5d2a08a754d)