video/drm: vop2: Use dev_read_addr_size_name() to get regs length for &vop2.regsbak allocation
If the DSC function is enabled on RK3588, there may be a crash due
to the out-of-bounds access to the &vop2.regsbak array. The reason
is that the size of &vop2.regsbak array is fixed to macro
RK3568_MAX_REG(0x1ed0) while the RK3588_DSC_8K_PPS0_3(0x4000) is out
of its range, which is writed via API vop2_writel().

Since different platform may have different regs length, we use API
dev_read_addr_size_name() to get the accurate size assigned in DTS.

In addtion, the regs size for backup is set to 0x2000 specifically
for RK3588, which is synchronized with the Kernel.

Fixes: d0408543457 ("drm/rockchip: Add support for vop2")
Change-Id: I4120fb6283e6a942b418033090685de25f07afa1
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>