video/drm: dw-dp: support more feature active protocol converter adapters
According to DP1.4a 5.3.3, To support more feature active protocol
converte adapters. All the adpters are brach device, Read the
Detailed Capabilities Info to get the branch type and the feature
it support. For Example, the max dot clock for VGA, max TDMS clock
for HDMI, the YCbCr420 support or not for HDMI 2.0, conversion to
YCbCr420 from YCbCr444 or not for HDMI 2.0.

According the feature of brach device, filter the resolution, color
format and color depth that the branch device can't support.

Signed-off-by: Zhang Yubing <yubing.zhang@rock-chips.com>
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>
Change-Id: Ib3ac1638d4ec8707dd66282d65133ddf3d22d6ee