video/drm: display: Use existing crc32() instead of custom crc32 related APIs
Apply crc32() API for crc calculation and remove custom APIs
rockchip_display_make_crc32_table() and rockchip_display_crc32c_cal().

In the past, the crc table was generated dynamically. Currently, it has
been modified to the same static table method as the application layer
for calculating the baseparemeter crc.

By the way, The crc results calculated by the two methods are the same,
only the mathematical implementations are different.

Change-Id: I3e7b5fa324cfd281a2a21932a1a6220f9538a458
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>