video/drm: vop2: Sync the default extra layer alpha configs with Kernel for rk3576
Add two new functions rk3576_extra_layer_sel_for_vp() and
rk3576_extra_alpha(). The former is to check whether there is extra
layer for VP0 and update layer select configurations. While the
latter is to update the extra layer alpha configurations.

They are both for synchronization with the Kernel.

Change-Id: I0fc9bb7d5af9c3bc76f1805a70a23dcdb35755f3
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>