video/drm: vop2: update color-encoding selection policy for post-csc
When all layers are rgb, the post-csc input color encoding
is fixed to rgb full, and the value of input_color_encoding
has no actual utility. If there are any yuv planes, value
of post-csc input_color_encoding selects the value of the
yuv plane with the largest area.

If all planes are in rgb format and are in yuv overlay, r2y
conversion is all carried out in bt601. In this scenario, the
input colorspace of post-csc needs to be selected as bt601.

Change-Id: I91f8f7338df9858bfb21e7e4d2dfb5268790b2b4
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>