video/drm: vop2: Optimize the lookup process for csc coef
Delete the cases where the input and output are exactly the
same (such as RGBF2RGBF), and uniformly replace them with the
identity matrix to reduce the repetitive parts in the table.

Change the original structure of large tables nested within
small tables into a single small table to enhance readability
of the code, also prepare for the increased csc accuracy for
subsequent platform such as rk3572.

Change-Id: I6ce33dce1f4c6c978867e51a35219260f61d1630
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>