video/drm: vop2: Fix failure of second csc table lookup
When the input/output has formats not supported by csc (such as the
conversion between bt2020 and bt709), the output colorspace will be
directly assigned to the input colorspace. After performing forced
assignment, if the input/output formats are exactly the same, csc
matrix of the IDENTITY_MODE needs to be used.

Change-Id: I03174affe5d263de341c23ed9104da03be0d33ba
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>