video/drm: vop2: Fix the specific address name to 'regs' for reading addr size
In the next-dev branch, the API for reading address size is
dev_read_addr_size(), which supports to get the address size of first
range in the 'reg' property.

For the for-v2025 branch, the previous API has become an inline
function. The similar API dev_read_addr_size_name() requires specific
address name, which assigned in the 'reg-names' property, as a
parameter.

Fixes: fea37e8db9c ("video/drm: vop2: Use dev_read_addr_size_name() to get regs length for &vop2.regsbak allocation")
Change-Id: I20299c8b3d23a91c951024c52e1d8eeacd67fa3f
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>