UPSTREAM: CVE-2025-24857: fs: fat: Perform sanity checks on getsize in get_fatent()
We do not perform a check on the value of getsize in get_fatent to
ensure that it will fit within the allocated buffer. For safety sake,
add a check now and if the value exceeds FATBUFBLOCKS use that value
instead. While not currently actively exploitable, it was in the past so
adding this check is worthwhile.

This addresses CVE-2025-24857 and was originally reported by Harvey
Phillips of Amazon Element55.

Signed-off-by: Tom Rini <trini@konsulko.com>
(cherry picked from commit 87d85139a96a39429120cca838e739408ef971a2)
Change-Id: I3ef4496ef9e5dd53d1607259e62b13a3b28ebbc0
Signed-off-by: Joseph Chen <chenjh@rock-chips.com>